+++ DragaliaAPI/DragaliaAPI.Database/ApiContext.cs
@@ -135,6 +135,10 @@ public class ApiContext : DbContext, IDataProtectionKeyContext
 
     public DbSet<DbPlayerHelper> PlayerHelpers { get; set; }
 
+    public DbSet<DbPlayerFriendship> PlayerFriendships { get; set; }
+
+    public DbSet<DbPlayerFriendshipPlayer> PlayerFriendshipPlayers { get; set; }
+
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
         modelBuilder.ApplyConfigurationsFromAssembly(typeof(ApiContext).Assembly);
