+++ DragaliaAPI/DragaliaAPI.Database/Entities/DbPlayer.cs
@@ -1,20 +1,18 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using Microsoft.AspNetCore.OutputCaching;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace DragaliaAPI.Database.Entities;
 
 /// <summary>
 /// Container class for all savefile data to enable foreign keys.
 /// </summary>
-[Index(nameof(AccountId))]
 public class DbPlayer
 {
-    [Key]
-    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public long ViewerId { get; set; }
 
-    [MaxLength(16)]
     public required string AccountId { get; set; }
 
     public DateTimeOffset CreatedAt { get; set; }
@@ -26,7 +24,6 @@ public class DbPlayer
     /// </summary>
     public DateTimeOffset? LastSavefileImportTime { get; set; }
 
-    [MaxLength(32)]
     public string? SavefileOrigin { get; set; }
 
     public bool IsAdmin { get; set; }
@@ -108,4 +105,19 @@ public class DbPlayer
     public DbPlayerDiamondData? DiamondData { get; set; }
 
     public DbPlayerHelper? Helper { get; set; }
+
+    public List<DbPlayerFriendship> Friendships { get; set; } = [];
+
+    private class Configuration : IEntityTypeConfiguration<DbPlayer>
+    {
+        public void Configure(EntityTypeBuilder<DbPlayer> builder)
+        {
+            builder.HasKey(e => e.ViewerId);
+
+            builder.HasIndex(e => e.AccountId);
+            builder.Property(e => e.AccountId).HasMaxLength(16);
+
+            builder.Property(e => e.SavefileOrigin).HasMaxLength(32);
+        }
+    }
 }
