+++ DragaliaAPI/DragaliaAPI.Database/Entities/DbPlayerFriendshipPlayer.cs
@@ -1,0 +1,15 @@
+namespace DragaliaAPI.Database.Entities;
+
+/// <summary>
+/// Join table for the many-to-many relationship between <see cref="DbPlayer"/> and <see cref="DbPlayerFriendship"/>.
+/// </summary>
+public class DbPlayerFriendshipPlayer
+{
+    public int PlayerId { get; set; }
+
+    public int PlayerFriendshipId { get; set; }
+
+    public DbPlayer? Player { get; set; }
+
+    public DbPlayerFriendship? Friendship { get; set; }
+}
