+++ DragaliaAPI/DragaliaAPI/Features/Friends/FriendService.cs
@@ -1,0 +1,44 @@
+using DragaliaAPI.Database;
+using DragaliaAPI.Database.Entities;
+using DragaliaAPI.Models.Generated;
+using DragaliaAPI.Shared.PlayerDetails;
+using Microsoft.EntityFrameworkCore;
+
+namespace DragaliaAPI.Features.Friends;
+
+internal sealed class FriendService(
+    ApiContext apiContext,
+    IPlayerIdentityService playerIdentityService
+)
+{
+    public async Task<int> GetFriendCount()
+    {
+        return await this.GetFriendsQuery().CountAsync();
+    }
+
+    public async Task<List<UserSupportList>> GetFriendList()
+    {
+        IQueryable<HelperProjection> helperQuery = this.GetFriendsQuery()
+            .Select(x => x.Helper!)
+            .ProjectToHelperProjection()
+            .AsSplitQuery();
+
+        List<HelperProjection> mergedHelpers = await helperQuery.ToListAsync();
+
+        return mergedHelpers.Select(x => x.MapToUserSupportList()).ToList();
+    }
+
+    private IQueryable<DbPlayer> GetFriendsQuery()
+    {
+        IQueryable<DbPlayer> currentPlayer = apiContext.Players.Where(x =>
+            x.ViewerId == playerIdentityService.ViewerId
+        );
+
+        return currentPlayer
+            .SelectMany(x => x.Friendships)
+            .Where(x => x.IsAccepted)
+            .SelectMany(x => x.Players)
+            .Where(x => x.ViewerId != playerIdentityService.ViewerId)
+            .IgnoreQueryFilters();
+    }
+}
